Refactor libs that query remote APIs
