+++ tasks/libs/common/gitlab.py
@@ -1,23 +1,29 @@
-import errno
 import json
 import os
 import platform
-import re
 import subprocess
 from urllib.parse import quote
 
 from invoke.exceptions import Exit
 
-errno_regex = re.compile(r".*\[Errno (\d+)\] (.*)")
+from .remote_api import RemoteAPI
 
 __all__ = ["Gitlab"]
 
 
-class Gitlab(object):
+class Gitlab(RemoteAPI):
     BASE_URL = "https://gitlab.ddbuild.io/api/v4"
 
     def __init__(self, api_token=None):
+        super(Gitlab, self).__init__()
+
         self.api_token = api_token if api_token else self._api_token()
+        self.api_name = "Gitlab"
+        self.authorization_error_message = (
+            "HTTP 401: Your GITLAB_TOKEN may have expired. You can "
+            "check and refresh it at "
+            "https://gitlab.ddbuild.io/profile/personal_access_tokens"
+        )
 
     def test_project_found(self, project):
         """
@@ -261,74 +267,23 @@ class Gitlab(object):
     ):
         """
         Utility to make a request to the Gitlab API.
+        See RemoteAPI#request.
 
-        headers: A hash of headers to pass to the request.
-        data: An object containing the body of the request.
-        json_input: If set to true, data is passed with the json parameter of requests.post instead of the data parameter.
-
-        By default, the request method is GET, or POST if data is not empty.
-        method: Can be set to "POST" to force a POST request even when data is empty.
-
-        By default, we return the text field of the response object. The following fields can alter this behavior:
-        json_output: the json field of the response object is returned.
-        stream_output: the request asks for a stream response, and the raw response object is returned.
+        Adds "PRIVATE-TOKEN: {self.api_token}" to the headers to be able to authenticate ourselves to GitLab.
         """
-        import requests
-
-        url = self.BASE_URL + path
-
         headers = dict(headers or [])
         headers["PRIVATE-TOKEN"] = self.api_token
 
-        # TODO: Use the param argument of requests instead of handling URL params
-        # manually
-        try:
-            # If json_input is true, we specifically want to send data using the json
-            # parameter of requests.post
-            if data and json_input:
-                r = requests.post(url, headers=headers, json=data, stream=stream_output)
-            elif method == "PUT":
-                r = requests.put(url, headers=headers, json=data, stream=stream_output)
-            elif method == "DELETE":
-                r = requests.delete(url, headers=headers, stream=stream_output)
-            elif data or method == "POST":
-                r = requests.post(url, headers=headers, data=data, stream=stream_output)
-            else:
-                r = requests.get(url, headers=headers, stream=stream_output)
-            if r.status_code == 401:
-                print(
-                    "HTTP 401: Your GITLAB_TOKEN may have expired. You can "
-                    "check and refresh it at "
-                    "https://gitlab.ddbuild.io/profile/personal_access_tokens"
-                )
-                print("Gitlab says: {}".format(r.json()["error_description"]))
-                raise Exit(code=1)
-        except requests.exceptions.Timeout:
-            print("Connection to GitLab ({}) timed out.".format(url))
-            raise Exit(code=1)
-        except requests.exceptions.RequestException as e:
-            m = errno_regex.match(str(e))
-            if not m:
-                print("Unknown error raised connecting to {}: {}".format(url, e))
-
-            # Parse errno to give a better explanation
-            # Requests doesn't have granularity at the level we want:
-            # http://docs.python-requests.org/en/master/_modules/requests/exceptions/
-            errno_code = int(m.group(1))
-            message = m.group(2)
-
-            if errno_code == errno.ENOEXEC:
-                print("Error resolving {}: {}".format(url, message))
-            elif errno_code == errno.ECONNREFUSED:
-                print("Connection to Gitlab ({}) refused".format(url))
-            else:
-                print("Error while connecting to {}: {}".format(url, str(e)))
-            raise Exit(code=1)
-        if json_output:
-            return r.json()
-        if stream_output:
-            return r
-        return r.text
+        return self.request(
+            path=path,
+            headers=headers,
+            data=data,
+            json_input=json_input,
+            json_output=json_output,
+            stream_output=stream_output,
+            raw_output=False,
+            method=method,
+        )
 
     def _api_token(self):
         if "GITLAB_TOKEN" not in os.environ:
