+++ tasks/libs/github_actions_tools.py
@@ -9,7 +9,7 @@ from invoke.exceptions import Exit
 from tasks.utils import DEFAULT_BRANCH
 
 from .common.color import color_message
-from .common.github import Github, GithubException
+from .common.github_workflows import GithubException, GithubWorkflows
 
 
 def trigger_macos_workflow(
@@ -47,7 +47,7 @@ def trigger_macos_workflow(
 
     # The workflow trigger endpoint doesn't return anything. You need to fetch the workflow run id
     # by yourself.
-    Github().trigger_workflow("DataDog/datadog-agent-macos-build", "macos.yaml", github_action_ref, inputs)
+    GithubWorkflows().trigger_workflow("DataDog/datadog-agent-macos-build", "macos.yaml", github_action_ref, inputs)
 
     # Thus the following hack: query the latest run for ref, wait until we get a non-completed run
     # that started after we triggered the workflow.
@@ -71,7 +71,9 @@ def get_macos_workflow_run_for_ref(github_action_ref="master"):
     """
     Get the latest workflow for the given ref.
     """
-    return Github().latest_workflow_run_for_ref("DataDog/datadog-agent-macos-build", "macos.yaml", github_action_ref)
+    return GithubWorkflows().latest_workflow_run_for_ref(
+        "DataDog/datadog-agent-macos-build", "macos.yaml", github_action_ref
+    )
 
 
 def follow_workflow_run(run_id):
@@ -80,7 +82,7 @@ def follow_workflow_run(run_id):
     """
 
     try:
-        run = Github().workflow_run("DataDog/datadog-agent-macos-build", run_id)
+        run = GithubWorkflows().workflow_run("DataDog/datadog-agent-macos-build", run_id)
     except GithubException:
         raise Exit(code=1)
 
@@ -96,7 +98,7 @@ def follow_workflow_run(run_id):
     while True:
         # Do not fail outright for temporary failures
         try:
-            run = Github().workflow_run("DataDog/datadog-agent-macos-build", run_id)
+            run = GithubWorkflows().workflow_run("DataDog/datadog-agent-macos-build", run_id)
         except GithubException:
             failures += 1
             print("Workflow run not found, retrying in 15 seconds (failure {}/{})".format(failures, MAX_FAILURES))
@@ -132,7 +134,7 @@ def download_artifacts(run_id, destination="."):
 
     print(color_message("Downloading artifacts for run {} to {}".format(run_id, destination), "blue"))
 
-    run_artifacts = Github().workflow_run_artifacts("DataDog/datadog-agent-macos-build", run_id)
+    run_artifacts = GithubWorkflows().workflow_run_artifacts("DataDog/datadog-agent-macos-build", run_id)
     if run_artifacts is None:
         print("Workflow run not found.")
         raise Exit(code=1)
@@ -141,7 +143,7 @@ def download_artifacts(run_id, destination="."):
     with tempfile.TemporaryDirectory() as tmpdir:
         for artifact in run_artifacts["artifacts"]:
             # Download artifact
-            zip_path = Github().download_artifact("DataDog/datadog-agent-macos-build", artifact["id"], tmpdir)
+            zip_path = GithubWorkflows().download_artifact("DataDog/datadog-agent-macos-build", artifact["id"], tmpdir)
 
             # Unzip it in the target destination
             with zipfile.ZipFile(zip_path, "r") as zip_ref:
